feat: add method for language switch and partly linter fix
