+++ .eslintrc
@@ -2,5 +2,8 @@
   "extends": ["airbnb-base"],
   "env": {
     "browser": true
+  },
+  "rules": {
+    "import/extensions": [ "error", "ignorePackages", { "": "never" } ]
   }
 }
