+++ index.html
@@ -11,7 +11,6 @@
   <title>Virtual-keyboard</title>
 </head>
 <body>
-  <script src="js/langs.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
