+++ js/components/textarea.js
@@ -1,0 +1,10 @@
+export class Textarea {
+  constructor() {
+    this.textarea = document.querySelector('.textarea');
+    document.addEventListener('keydown', this.onKeydown);
+  }
+
+  onKeydown = () => {
+    this.textarea.focus();
+  }
+}
