+++ js/keySets.js
@@ -1,0 +1,647 @@
+export const keySets = {
+  en: [[{
+    value: '`',
+    shiftValue: '~',
+    keycode: 'backquote',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '1',
+    shiftValue: '!',
+    keycode: 'digit1',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '2',
+    shiftValue: '@',
+    keycode: 'digit2',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '3',
+    shiftValue: '#',
+    keycode: 'digit3',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '4',
+    shiftValue: '$',
+    keycode: 'digit4',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '5',
+    shiftValue: '%',
+    keycode: 'digit5',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '6',
+    shiftValue: '^',
+    keycode: 'digit6',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '7',
+    shiftValue: '&',
+    keycode: 'digit7',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '8',
+    shiftValue: '*',
+    keycode: 'digit8',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '9',
+    shiftValue: '(',
+    keycode: 'digit9',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '0',
+    shiftValue: ')',
+    keycode: 'digit0',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '-',
+    shiftValue: '_',
+    keycode: 'minus',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '=',
+    shiftValue: '+',
+    keycode: 'equal',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: 'Backspace',
+    shiftValue: 'Backspace',
+    keycode: 'backspace',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Tab',
+    shiftValue: 'Tab',
+    keycode: 'tab',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'q',
+    shiftValue: 'Q',
+    keycode: 'keyq',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'w',
+    shiftValue: 'W',
+    keycode: 'keyw',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'e',
+    shiftValue: 'E',
+    keycode: 'keye',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'r',
+    shiftValue: 'R',
+    keycode: 'keyr',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 't',
+    shiftValue: 'T',
+    keycode: 'keyt',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'y',
+    shiftValue: 'Y',
+    keycode: 'keyy',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'u',
+    shiftValue: 'U',
+    keycode: 'keyu',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'i',
+    shiftValue: 'I',
+    keycode: 'keyi',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'o',
+    shiftValue: 'O',
+    keycode: 'keyo',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'p',
+    shiftValue: 'P',
+    keycode: 'keyp',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: '[',
+    shiftValue: '{',
+    keycode: 'bracketleft',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: ']',
+    shiftValue: '}',
+    keycode: 'bracketright',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '\\',
+    shiftValue: '|',
+    keycode: 'backslash',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: 'Del',
+    shiftValue: 'Del',
+    keycode: 'delete',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }], [{
+    value: 'CapsLock',
+    shiftValue: 'CapsLock',
+    keycode: 'capslock',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }, {
+    value: 'a',
+    shiftValue: 'A',
+    keycode: 'keya',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 's',
+    shiftValue: 'S',
+    keycode: 'keys',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'd',
+    shiftValue: 'D',
+    keycode: 'keyd',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'f',
+    shiftValue: 'F',
+    keycode: 'keyf',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'g',
+    shiftValue: 'G',
+    keycode: 'keyg',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'h',
+    shiftValue: 'H',
+    keycode: 'keyh',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'j',
+    shiftValue: 'J',
+    keycode: 'keyj',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'k',
+    shiftValue: 'K',
+    keycode: 'keyk',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'l',
+    shiftValue: 'L',
+    keycode: 'keyl',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: ';',
+    shiftValue: ':',
+    keycode: 'semicolon',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '\'',
+    shiftValue: '"',
+    keycode: 'quote',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: 'Enter',
+    shiftValue: 'Enter',
+    keycode: 'enter',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Shift',
+    shiftValue: 'Shift',
+    keycode: 'shiftleft',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }, {
+    value: 'z',
+    shiftValue: 'Z',
+    keycode: 'keyz',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'x',
+    shiftValue: 'X',
+    keycode: 'keyx',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'c',
+    shiftValue: 'C',
+    keycode: 'keyc',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'v',
+    shiftValue: 'V',
+    keycode: 'keyv',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'b',
+    shiftValue: 'B',
+    keycode: 'keyb',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'n',
+    shiftValue: 'N',
+    keycode: 'keyn',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'm',
+    shiftValue: 'M',
+    keycode: 'keym',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: ',',
+    shiftValue: '<',
+    keycode: 'comma',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '.',
+    shiftValue: '>',
+    keycode: 'period',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '/',
+    shiftValue: '?',
+    keycode: 'slash',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '↑',
+    shiftValue: '↑',
+    keycode: 'arrowup',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: 'Shift',
+    shiftValue: 'Shift',
+    keycode: 'shiftright',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Ctrl',
+    shiftValue: 'Ctrl',
+    keycode: 'controlleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'Win',
+    shiftValue: 'Win',
+    keycode: 'metaleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'Alt',
+    shiftValue: 'Alt',
+    keycode: 'altleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: ' ',
+    shiftValue: ' ',
+    keycode: 'space',
+    class: 'keyboard__btn keyboard__btn_space',
+  }, {
+    value: 'Alt',
+    shiftValue: 'Alt',
+    keycode: 'altright',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: '←',
+    shiftValue: '←',
+    keycode: 'arrowleft',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: '↓',
+    shiftValue: '↓',
+    keycode: 'arrowdown',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: '→',
+    shiftValue: '→',
+    keycode: 'arrowright',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: 'Ctrl',
+    shiftValue: 'Ctrl',
+    keycode: 'controlright',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }]],
+  ru: [[{
+    value: 'ё',
+    shiftValue: 'Ё',
+    keycode: 'backquote',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: '1',
+    shiftValue: '!',
+    keycode: 'digit1',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '2',
+    shiftValue: '"',
+    keycode: 'digit2',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '3',
+    shiftValue: '№',
+    keycode: 'digit3',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '4',
+    shiftValue: ';',
+    keycode: 'digit4',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '5',
+    shiftValue: '%',
+    keycode: 'digit5',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '6',
+    shiftValue: ':',
+    keycode: 'digit6',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '7',
+    shiftValue: '?',
+    keycode: 'digit7',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '8',
+    shiftValue: '*',
+    keycode: 'digit8',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '9',
+    shiftValue: '(',
+    keycode: 'digit9',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '0',
+    shiftValue: ')',
+    keycode: 'digit0',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '-',
+    shiftValue: '_',
+    keycode: 'minus',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '=',
+    shiftValue: '+',
+    keycode: 'equal',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: 'Backspace',
+    shiftValue: 'Backspace',
+    keycode: 'backspace',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Tab',
+    shiftValue: 'Tab',
+    keycode: 'tab',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'й',
+    shiftValue: 'Й',
+    keycode: 'keyq',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ц',
+    shiftValue: 'Ц',
+    keycode: 'keyw',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'у',
+    shiftValue: 'У',
+    keycode: 'keye',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'к',
+    shiftValue: 'К',
+    keycode: 'keyr',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'е',
+    shiftValue: 'Е',
+    keycode: 'keyt',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'н',
+    shiftValue: 'Н',
+    keycode: 'keyy',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'г',
+    shiftValue: 'Г',
+    keycode: 'keyu',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ш',
+    shiftValue: 'Ш',
+    keycode: 'keyi',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'щ',
+    shiftValue: 'Щ',
+    keycode: 'keyo',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'з',
+    shiftValue: 'З',
+    keycode: 'keyp',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'х',
+    shiftValue: 'Х',
+    keycode: 'bracketleft',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ъ',
+    shiftValue: 'Ъ',
+    keycode: 'bracketright',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: '\\',
+    shiftValue: '/',
+    keycode: 'backslash',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: 'Del',
+    shiftValue: 'Del',
+    keycode: 'delete',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }], [{
+    value: 'CapsLock',
+    shiftValue: 'CapsLock',
+    keycode: 'capslock',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }, {
+    value: 'ф',
+    shiftValue: 'Ф',
+    keycode: 'keya',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ы',
+    shiftValue: 'Ы',
+    keycode: 'keys',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'в',
+    shiftValue: 'В',
+    keycode: 'keyd',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'а',
+    shiftValue: 'А',
+    keycode: 'keyf',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'п',
+    shiftValue: 'П',
+    keycode: 'keyg',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'р',
+    shiftValue: 'Р',
+    keycode: 'keyh',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'о',
+    shiftValue: 'О',
+    keycode: 'keyj',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'л',
+    shiftValue: 'Л',
+    keycode: 'keyk',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'д',
+    shiftValue: 'Д',
+    keycode: 'keyl',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ж',
+    shiftValue: 'Ж',
+    keycode: 'semicolon',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'э',
+    shiftValue: 'Э',
+    keycode: 'quote',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'Enter',
+    shiftValue: 'Enter',
+    keycode: 'enter',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Shift',
+    shiftValue: 'Shift',
+    keycode: 'shiftleft',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }, {
+    value: 'я',
+    shiftValue: 'Я',
+    keycode: 'keyz',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ч',
+    shiftValue: 'Ч',
+    keycode: 'keyx',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'c',
+    shiftValue: 'С',
+    keycode: 'keyc',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'м',
+    shiftValue: 'М',
+    keycode: 'keyv',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'и',
+    shiftValue: 'И',
+    keycode: 'keyb',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'т',
+    shiftValue: 'Т',
+    keycode: 'keyn',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ь',
+    shiftValue: 'Ь',
+    keycode: 'keym',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'б',
+    shiftValue: 'Б',
+    keycode: 'comma',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: 'ю',
+    shiftValue: 'Ю',
+    keycode: 'period',
+    class: 'keyboard__btn keyboard__btn_letter',
+  }, {
+    value: '.',
+    shiftValue: ',',
+    keycode: 'slash',
+    class: 'keyboard__btn keyboard__btn_symbol',
+  }, {
+    value: '↑',
+    shiftValue: '↑',
+    keycode: 'arrowup',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: 'Shift',
+    shiftValue: 'Shift',
+    keycode: 'shiftright',
+    class: 'keyboard__btn keyboard__btn_comand_big',
+  }], [{
+    value: 'Ctrl',
+    shiftValue: 'Ctrl',
+    keycode: 'controlleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'Win',
+    shiftValue: 'Win',
+    keycode: 'metaleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: 'Alt',
+    shiftValue: 'Alt',
+    keycode: 'altleft',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: ' ',
+    shiftValue: ' ',
+    keycode: 'space',
+    class: 'keyboard__btn keyboard__btn_space',
+  }, {
+    value: 'Alt',
+    shiftValue: 'Alt',
+    keycode: 'altright',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }, {
+    value: '←',
+    shiftValue: '←',
+    keycode: 'arrowleft',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: '↓',
+    shiftValue: '↓',
+    keycode: 'arrowdown',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: '→',
+    shiftValue: '→',
+    keycode: 'arrowright',
+    class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
+  }, {
+    value: 'Ctrl',
+    shiftValue: 'Ctrl',
+    keycode: 'controlright',
+    class: 'keyboard__btn keyboard__btn_comand_small',
+  }]],
+};
+
+export const LANG_EN = 'en';
+export const LANG_RU = 'ru';
