--- js/langs.js
@@ -1,578 +1,0 @@
-const en = [[{
-  value: '`',
-  shiftValue: '~',
-  keycode: 'backquote',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '1',
-  shiftValue: '!',
-  keycode: 'digit1',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '2',
-  shiftValue: '@',
-  keycode: 'digit2',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '3',
-  shiftValue: '#',
-  keycode: 'digit3',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '4',
-  shiftValue: '$',
-  keycode: 'digit4',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '5',
-  shiftValue: '%',
-  keycode: 'digit5',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '6',
-  shiftValue: '^',
-  keycode: 'digit6',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '7',
-  shiftValue: '&',
-  keycode: 'digit7',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '8',
-  shiftValue: '*',
-  keycode: 'digit8',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '9',
-  shiftValue: '(',
-  keycode: 'digit9',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '0',
-  shiftValue: ')',
-  keycode: 'digit0',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '-',
-  shiftValue: '_',
-  keycode: 'minus',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '=',
-  shiftValue: '+',
-  keycode: 'equal',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: 'Backspace',
-  shiftValue: 'Backspace',
-  keycode: 'backspace',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Tab',
-  shiftValue: 'Tab',
-  keycode: 'tab',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'q',
-  shiftValue: 'Q',
-  keycode: 'keyq',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'w',
-  shiftValue: 'W',
-  keycode: 'keyw',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'e',
-  shiftValue: 'E',
-  keycode: 'keye',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'r',
-  shiftValue: 'R',
-  keycode: 'keyr',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 't',
-  shiftValue: 'T',
-  keycode: 'keyt',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'y',
-  shiftValue: 'Y',
-  keycode: 'keyy',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'u',
-  shiftValue: 'U',
-  keycode: 'keyu',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'i',
-  shiftValue: 'I',
-  keycode: 'keyi',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'o',
-  shiftValue: 'O',
-  keycode: 'keyo',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'p',
-  shiftValue: 'P',
-  keycode: 'keyp',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: '[',
-  shiftValue: '{',
-  keycode: 'bracketleft',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: ']',
-  shiftValue: '}',
-  keycode: 'bracketright',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '\\',
-  shiftValue: '|',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: 'Del',
-  shiftValue: 'Del',
-  keycode: 'delete',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}], [{
-  value: 'CapsLock',
-  shiftValue: 'CapsLock',
-  keycode: 'capslock',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}, {
-  value: 'a',
-  shiftValue: 'A',
-  keycode: 'keya',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 's',
-  shiftValue: 'S',
-  keycode: 'keys',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'd',
-  shiftValue: 'D',
-  keycode: 'keyd',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'f',
-  shiftValue: 'F',
-  keycode: 'keyf',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'g',
-  shiftValue: 'G',
-  keycode: 'keyg',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'h',
-  shiftValue: 'H',
-  keycode: 'keyh',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'j',
-  shiftValue: 'J',
-  keycode: 'keyj',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'k',
-  shiftValue: 'K',
-  keycode: 'keyk',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'l',
-  shiftValue: 'L',
-  keycode: 'keyl',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: ';',
-  shiftValue: ':',
-  keycode: 'semicolon',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '\'',
-  shiftValue: '"',
-  keycode: 'quote',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: 'Enter',
-  shiftValue: 'Enter',
-  keycode: 'enter',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Shift',
-  shiftValue: 'Shift',
-  keycode: 'shiftleft',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}, {
-  value: 'z',
-  shiftValue: 'Z',
-  keycode: 'keyz',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'x',
-  shiftValue: 'X',
-  keycode: 'keyx',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'c',
-  shiftValue: 'C',
-  keycode: 'keyc',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'v',
-  shiftValue: 'V',
-  keycode: 'keyv',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'b',
-  shiftValue: 'B',
-  keycode: 'keyb',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'n',
-  shiftValue: 'N',
-  keycode: 'keyn',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'm',
-  shiftValue: 'M',
-  keycode: 'keym',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: ',',
-  shiftValue: '<',
-  keycode: 'comma',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '.',
-  shiftValue: '>',
-  keycode: 'period',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '/',
-  shiftValue: '?',
-  keycode: 'slash',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '↑',
-  shiftValue: '↑',
-  keycode: 'arrowup',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: 'Shift',
-  shiftValue: 'Shift',
-  keycode: 'shiftright',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Ctrl',
-  shiftValue: 'Ctrl',
-  keycode: 'controlleft',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'Win',
-  shiftValue: 'Win',
-  keycode: 'metaleft',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'Alt',
-  shiftValue: 'Alt',
-  keycode: 'altleft',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: ' ',
-  shiftValue: ' ',
-  keycode: 'space',
-  class: 'keyboard__btn keyboard__btn_space',
-}, {
-  value: 'Alt',
-  shiftValue: 'Alt',
-  keycode: 'altright',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: '←',
-  shiftValue: '←',
-  keycode: 'arrowleft',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: '↓',
-  shiftValue: '↓',
-  keycode: 'arrowdown',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: '→',
-  shiftValue: '→',
-  keycode: 'arrowright',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: 'Ctrl',
-  shiftValue: 'Ctrl',
-  keycode: 'controlright',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}]];
-
-const ru = [[{
-  value: 'ё',
-  shiftValue: 'Ё',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: '1',
-  shiftValue: '!',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '2',
-  shiftValue: '"',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '3',
-  shiftValue: '№',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '4',
-  shiftValue: ';',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '5',
-  shiftValue: '%',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '6',
-  shiftValue: ':',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '7',
-  shiftValue: '?',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '8',
-  shiftValue: '*',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '9',
-  shiftValue: '(',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '0',
-  shiftValue: ')',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '-',
-  shiftValue: '_',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '=',
-  shiftValue: '+',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: 'Backspace',
-  shiftValue: 'Backspace',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Tab',
-  shiftValue: 'Tab',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'й',
-  shiftValue: 'Й',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ц',
-  shiftValue: 'Ц',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'у',
-  shiftValue: 'У',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'к',
-  shiftValue: 'К',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'е',
-  shiftValue: 'Е',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'н',
-  shiftValue: 'Н',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'г',
-  shiftValue: 'Г',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ш',
-  shiftValue: 'Ш',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'щ',
-  shiftValue: 'Щ',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'з',
-  shiftValue: 'З',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'х',
-  shiftValue: 'Х',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ъ',
-  shiftValue: 'Ъ',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: '\\',
-  shiftValue: '/',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: 'Del',
-  shiftValue: 'Del',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}], [{
-  value: 'CapsLock',
-  shiftValue: 'CapsLock',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}, {
-  value: 'ф',
-  shiftValue: 'Ф',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ы',
-  shiftValue: 'Ы',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'в',
-  shiftValue: 'В',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'а',
-  shiftValue: 'А',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'п',
-  shiftValue: 'П',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'р',
-  shiftValue: 'Р',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'о',
-  shiftValue: 'О',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'л',
-  shiftValue: 'Л',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'д',
-  shiftValue: 'Д',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ж',
-  shiftValue: 'Ж',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'э',
-  shiftValue: 'Э',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'Enter',
-  shiftValue: 'Enter',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Shift',
-  shiftValue: 'Shift',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}, {
-  value: 'я',
-  shiftValue: 'Я',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ч',
-  shiftValue: 'Ч',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'c',
-  shiftValue: 'С',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'м',
-  shiftValue: 'М',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'и',
-  shiftValue: 'И',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'т',
-  shiftValue: 'Т',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ь',
-  shiftValue: 'Ь',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'б',
-  shiftValue: 'Б',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: 'ю',
-  shiftValue: 'Ю',
-  class: 'keyboard__btn keyboard__btn_letter',
-}, {
-  value: '.',
-  shiftValue: ',',
-  class: 'keyboard__btn keyboard__btn_symbol',
-}, {
-  value: '↑',
-  shiftValue: '↑',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: 'Shift',
-  shiftValue: 'Shift',
-  class: 'keyboard__btn keyboard__btn_comand_big',
-}], [{
-  value: 'Ctrl',
-  shiftValue: 'Ctrl',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'Win',
-  shiftValue: 'Win',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: 'Alt',
-  shiftValue: 'Alt',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: ' ',
-  shiftValue: ' ',
-  class: 'keyboard__btn keyboard__btn_space',
-}, {
-  value: 'Alt',
-  shiftValue: 'Alt',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}, {
-  value: '←',
-  shiftValue: '←',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: '↓',
-  shiftValue: '↓',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: '→',
-  shiftValue: '→',
-  class: 'keyboard__btn keyboard__btn_comand_small keyboard__btn_comand_arrow',
-}, {
-  value: 'Ctrl',
-  shiftValue: 'Ctrl',
-  class: 'keyboard__btn keyboard__btn_comand_small',
-}]];
