+++ package.json
@@ -11,6 +11,8 @@
   "author": "elen-jagger",
   "license": "ISC",
   "devDependencies": {
+    "eslint": "^8.39.0",
+    "eslint-cli": "^1.1.1",
     "eslint-config-airbnb-base": "^15.0.0"
   }
 }
